created class `FrameworkBase`
