+++ doc/userguide/api/framework.md
@@ -4,5 +4,6 @@
 ```{eval-rst}
 .. automodule:: pyrigi.framework
    :members:
+   :inherited-members:
    :show-inheritance:
 ```
